// src/orders/enums/order-status.enum.ts
export enum OrderStatus {
    EXPIRED_BANK_SLIP = 'EXPIRED_BANK_SLIP',
    FAILED_PAYMENT = 'FAILED_PAYMENT',
    REFUNDED_PAYMENT = 'REFUNDED_PAYMENT',
    PENDING_PAYMENT = 'PENDING_PAYMENT',
    PENDING_SHIPPING_LABELS = 'PENDING_SHIPPING_LABELS',
    PENDING_SHIPPING = 'PENDING_SHIPPING',
    PENDING_DELIVERY = 'PENDING_DELIVERY',
    FINISHED = 'FINISHED',
    CANCELED = 'CANCELED'
}